Mesos module APIs moved to separate folder
