+++ src/mesos_module/resource_estimator/serenity_estimator.cpp
@@ -1,9 +1,9 @@
 #include <list>
 #include <memory>
 
-#include "estimator/serenity_estimator.hpp"
-
 #include "glog/logging.h"
+
+#include "mesos_module/resource_estimator/serenity_estimator.hpp"
 
 #include "pipeline/estimator_pipeline.hpp"
