+++ src/mesos_module/resource_estimator/serenity_estimator_module.cpp
@@ -1,13 +1,13 @@
 #include <memory>
 #include <string>
 
-#include "estimator/serenity_estimator.hpp"
-
 #include "mesos/mesos.hpp"
 #include "mesos/module.hpp"
 
 #include "mesos/module/resource_estimator.hpp"
 #include "mesos/slave/resource_estimator.hpp"
+
+#include "mesos_module/resource_estimator/serenity_estimator.hpp"
 
 #include "pipeline/estimator_pipeline.hpp"
